pyproject.toml
[project]
name = "de-test-junior"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama>=0.4.6",
    "cryptography>=45.0.6",
    "faker>=37.5.3",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]
